# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zonvo AI REST API
  description: |
    Welcome to the Zonvo AI REST API. Everything your AI voice agents do on the phone — calling, qualifying,
    booking, promising a callback, honouring a do-not-call request — is available to your CRM, ERP or
    in-house app through `https://app.zonvo.tech/api/v1`.

    ## What you can automate

    | Goal | Endpoints |
    |------|-----------|
    | **Trigger calls** from your own app, now or at a scheduled time | `POST /calls` (with `scheduledAt`, `dynamicVariables`) |
    | **Run bulk campaigns** | `POST /campaigns`, `/campaigns/{id}/contacts`, `/start`, `/pause` |
    | **Sync leads and call outcomes to your CRM** | `GET /leads`, `PATCH /leads/{id}`, `POST /leads/{id}/notes`, `GET /calls?outcome=`, `POST /calls/{id}/outcome` |
    | **Book, confirm or cancel appointments** the agent made | `GET /appointments`, `PATCH /appointments/{id}` |
    | **Schedule callbacks** (single or in bulk, with idempotency keys) | `POST /callbacks`, `POST /callbacks/bulk`, `DELETE /callbacks?externalRef=` |
    | **Keep a do-not-call list** in sync with your compliance system | `GET /dnd`, `POST /dnd`, `POST /dnd/import`, `GET /dnd/check` |
    | **Receive real-time webhooks** for calls, appointments, callbacks and leads | `POST /webhooks`, `GET /webhooks/events` |

    ## Integrating from a CRM in 5 steps

    1. **Create an API key** — Settings → Developer → *Create API Key*. Give it only the scopes you need
...

...

...

...

...
        ]
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /calls:
    get:
      summary: List Calls
      description: |
        Retrieve a paginated list of all calls made from your account. Results include calls from all telephony engines (ElevenLabs, Plivo, Twilio-OpenAI).
        
        **Tip:** Use the `page` and `pageSize` parameters to navigate through large result sets efficiently.
        
        **Required Scope:** `calls:read`
      operationId: listCalls
      tags: [Calls]
      parameters:
        - name: outcome
          in: query
          description: Only calls with this outcome (see `GET /calls/outcomes`; unknown ids return 400)
          schema:
            $ref: '#/components/schemas/CallOutcome'
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          description: Calls created at or after this time (ISO 8601)
          schema:
            type: string
            format: date-time
          example: "2026-09-01T00:00:00+05:30"
        - name: to
          in: query
          description: Calls created before this time (ISO 8601)
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of calls with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Trigger Call
      description: |
        Initiate an outbound call using a specified AI agent. The call will be placed immediately unless scheduled for a future time.
        
        <br />
        
        **Prerequisites:**
        - An active AI agent (flow or incoming type)
        - Sufficient credits in your account
        - An active phone number (or provide `fromNumber`)
        - The number is not on your do-not-call list (`GET /dnd/check`)

        <br />

        **Scheduling:** pass a future `scheduledAt` and the call is queued as a callback instead of being
        placed now — the response is `202` with `callbackId`; manage it through the Callbacks endpoints.

        **Personalisation:** `dynamicVariables` (≤ 20 keys, names `[A-Za-z0-9_]{1,40}`, values ≤ 300 chars) are
        substituted into the agent's prompt and first message as `{{key}}` (or `{key}`) on every engine, and listed
        to the agent as *Details for this call*. Put your own record id in `metadata.externalRef`: it is stored on
        the call, echoed in `call.completed`, and — for scheduled calls — used as the idempotency key.
        
        <br />
        
        **Telephony Engines:**
        - `elevenlabs` — Best voice quality, 29 languages
        - `plivo` — India/APAC optimized
        - `twilio-openai` — Global coverage
        
        <br />
        
        **Required Scope:** `calls:write`
      operationId: triggerCall
      tags: [Calls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerCallRequest'
            examples:
              basic:
                summary: Basic call
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+14155551234"
              withEngine:
                summary: Specify engine
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+14155551234"
                  engine: "twilio-openai"
                  fromNumber: "+14155559999"
              personalised:
                summary: CRM lead with dynamic variables
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+919876543210"
                  dynamicVariables:
                    first_name: "Priya"
                    order_id: "ORD-48213"
                    delivery_date: "26 September"
                  metadata:
                    crmId: "SF-00412"
              scheduled:
                summary: Schedule for later (returns 202)
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+919876543210"
                  scheduledAt: "2026-09-29T11:00:00+05:30"
                  dynamicVariables:
                    first_name: "Priya"
      responses:
        '201':
          description: Call initiated successfully. Returns call ID and initial status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerCallResponse'
        '202':
          description: Call scheduled for `scheduledAt` (a callback was created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledCallResponse'
              example:
                success: true
                data:
                  scheduled: true
                  callbackId: "7d1a6c9e-0b2f-4e3a-9c8d-5f6a7b8c9d0e"
                  scheduledAt: "2026-09-29T05:30:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Number is on the do-not-call list (`error.code` = `DO_NOT_CALL`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /calls/{id}:
    get:
      summary: Get Call Details
      description: |
        Retrieve detailed information about a specific call including:
        - Call status and duration
        - Full conversation transcript
        - AI-generated summary
        - Sentiment analysis
        - Recording URL (if available)
        
        <br />
        
        **Tip:** Poll this endpoint to track call progress. Status will transition from `queued` → `ringing` → `in-progress` → `completed`.
        
        <br />
        
        **Required Scope:** `calls:read`
      operationId: getCall
      tags: [Calls]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Complete call details with transcript and analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /analytics/calls:
    get:
      summary: Get Call Analytics
      description: |
        Retrieve aggregated call statistics. Useful for dashboards and performance monitoring.
        
        <br />
        
        **Metrics Included:**
        - Total calls and minutes
        - Answer rate and average duration
        - Outcome distribution (completed, failed, no-answer)
        - Daily trends
        
        <br />
        
        **Required Scope:** `analytics:read`
      operationId: getCallAnalytics
      tags: [Analytics]
      parameters:
        - name: days
          in: query
          description: Number of days to analyze (default 30, max 90)
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: Call analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /webhooks/{id}/test:
    post:
      summary: Test Webhook
      description: |
        Send a test payload to verify your webhook endpoint is configured correctly. The test payload mimics a real event.
        
        <br />
        
        **Test Process:**
        1. Zonvo AI sends a test payload to your URL
        2. Your server should return 2xx status
        3. Response includes success/failure and timing info
        
        <br />
        
        **Troubleshooting:**
        - Ensure your endpoint is publicly accessible
        - Check that you respond within 30 seconds
        - Verify SSL certificate is valid
        
        <br />
        
        **Required Scope:** `webhooks:write`
      operationId: testWebhook
      tags: [Webhooks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Test results with timing and response info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookTestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /callbacks:
    get:
      summary: List Callbacks
      description: |
        Scheduled, running and finished callbacks, newest first. An unknown `status` value returns 400.

        **Required Scope:** `calls:read`
      operationId: listCallbacks
      tags: [Callbacks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, calling, completed, failed, cancelled]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Callbacks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackListResponse'
              example:
                success: true
                data:
                  - id: "7d1a6c9e-0b2f-4e3a-9c8d-5f6a7b8c9d0e"
                    contactName: "Priya Sharma"
                    contactPhone: "+919876543210"
                    reason: "Webinar reminder (1 day before)"
                    scheduledAt: "2026-09-29T05:30:00.000Z"
                    timeZone: "Asia/Kolkata"
                    status: pending
                    attempts: 0
                    lastError: null
                    agentId: "550e8400-e29b-41d4-a716-446655440000"
                    agentName: "Webinar reminder"
                    resultCallId: null
                    variables: { first_name: "Priya", webinar_title: "GST filing for startups" }
                    externalRef: "reg_8812-24h"
                    source: api
                    createdAt: "2026-09-24T08:00:00.000Z"
                meta:
                  requestId: "6b1f0c1e-3d2a-4c5b-8e9f-0a1b2c3d4e5f"
                  timestamp: "2026-09-24T08:00:00.000Z"
                  pagination: { page: 1, pageSize: 20, totalItems: 1, totalPages: 1, hasNext: false, hasPrev: false }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Schedule a Callback
      description: |
        Schedule one automatic call. The agent must belong to you; `scheduledAt` must be in the future and at most
        60 days ahead; the number must not be on the do-not-call list.

        Pass `externalRef` to make the request idempotent — the same ref returns the existing callback with
        HTTP 200. Pass `variables` to personalise the conversation (`{{key}}` in the prompt / first message).

        **Required Scope:** `calls:write`
      operationId: createCallback
      tags: [Callbacks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCallbackRequest'
            example:
              agentId: "550e8400-e29b-41d4-a716-446655440000"
              contactPhone: "+919876543210"
              contactName: "Priya Sharma"
              scheduledAt: "2026-09-29T11:00:00+05:30"
              timeZone: "Asia/Kolkata"
              reason: "Follow up on the pricing question"
              externalRef: "crm-task-48213"
              variables: { first_name: "Priya", plan: "Growth" }
      responses:
        '201':
          description: Callback scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponse'
        '200':
          description: An identical `externalRef` already exists — the existing callback is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponse'
        '400':
          description: Validation error — `error.details.reason` is `past`, `too_far` or `invalid_phone` when applicable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Number is on the do-not-call list (`error.code` = `DO_NOT_CALL`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: { code: DO_NOT_CALL, message: "This number is on your do-not-call list.", details: { reason: do_not_call } }
    delete:
      summary: Cancel a Callback by External Reference
      description: |
        Cancel the pending callback that was created with this `externalRef`.

        **Required Scope:** `calls:write`
      operationId: cancelCallbackByRef
      tags: [Callbacks]
      parameters:
        - name: externalRef
          in: query
          required: true
          schema:
            type: string
            maxLength: 120
          example: "reg_8812-24h"
      responses:
        '200':
          description: Callback cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackCancelResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /callbacks/{id}:
    delete:
      summary: Cancel a Callback
      description: |
        Cancel a pending callback. Returns 404 when it does not exist, is not yours, or is no longer pending.

        **Required Scope:** `calls:write`
      operationId: cancelCallback
      tags: [Callbacks]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Callback cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackCancelResponse'
              example:
                success: true
                data:
                  cancelled: true
                  callback: { id: "7d1a6c9e-0b2f-4e3a-9c8d-5f6a7b8c9d0e", contactPhone: "+919876543210", status: cancelled, externalRef: "reg_8812-24h", source: api }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InsufficientCredits:
      description: Insufficient credits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                hasNext:
                  type: boolean
                hasPrev:
                  type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
        meta:
          $ref: '#/components/schemas/Meta'
...
    TriggerCallRequest:
      type: object
      required:
        - agentId
        - toNumber
      properties:
        agentId:
          type: string
          format: uuid
          description: ID of the agent to use for the call
        toNumber:
          type: string
          description: Phone number to call (E.164 format recommended)
          example: "+14155551234"
        fromNumber:
          type: string
          description: Caller ID phone number (optional, uses default if not provided)
        engine:
          type: string
          enum: [elevenlabs, plivo, twilio-openai]
          description: Telephony engine to use
        dynamicVariables:
          type: object
          additionalProperties:
            type: string
          description: Substituted into the agent's prompt / first message as `{{key}}`. Max 20 keys, key `[A-Za-z0-9_]{1,40}`, values ≤ 300 chars.
          maxProperties: 20
          example: { first_name: "Priya", order_id: "ORD-48213" }
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Custom metadata to attach to the call (returned in webhooks as `data.metadata`, e.g. your CRM id or `externalRef`)
        scheduledAt:
          type: string
          format: date-time
          description: Future time (≤ 60 days) — the call is scheduled as a callback and the response is 202
    TriggerCallResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            callId:
              type: string
              format: uuid
            status:
              type: string
              example: queued
            agentId:
              type: string
              format: uuid
            toNumber:
              type: string
            fromNumber:
              type: string
            engine:
              type: string
            createdAt:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
    Call:
      type: object
      properties:
        id:
          type: string
          format: uuid
        engine:
          type: string
          enum: [elevenlabs, plivo, twilio-openai]
        agentId:
          type: string
          format: uuid
        toNumber:
          type: string
        fromNumber:
          type: string
        status:
          type: string
          enum: [queued, ringing, in-progress, completed, failed, no-answer, busy]
        duration:
          type: integer
          description: Duration in seconds
        creditsUsed:
          type: number
          nullable: true
          description: null for plivo / twilio-openai calls (minutes are billed from duration)
        campaignId:
          type: string
          format: uuid
          nullable: true
        outcome:
          allOf:
            - $ref: '#/components/schemas/CallOutcome'
          nullable: true
          description: null while the call is in progress or when nothing was recorded
        outcomeSource:
          type: string
          nullable: true
          enum: [agent, system, ai, null]
        answeredBy:
          type: string
          nullable: true
          enum: [human, machine, null]
          description: From answering-machine detection on outbound calls
        transcript:
          type: string
        aiSummary:
          type: string
        recordingUrl:
          type: string
        sentiment:
          type: string
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    CallListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Call'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CallDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Call'
        meta:
          $ref: '#/components/schemas/Meta'
...
    CallAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalCalls:
              type: integer
            completedCalls:
              type: integer
            failedCalls:
              type: integer
            totalDurationMinutes:
              type: integer
            averageDurationSeconds:
              type: integer
            creditsUsed:
              type: number
            period:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
...
    WebhookTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            delivered:
              type: boolean
            statusCode:
              type: integer
            message:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'

    # ---------------------------------------------------------------- Call outcomes
    CallOutcome:
      type: string
      description: How the call went. Agent-kind outcomes are set in conversation or via the API; system-kind by the platform.
      enum: [interested, not_interested, callback_requested, wrong_number, already_customer, do_not_call, no_decision, voicemail, no_answer, busy, failed, transferred, appointment_booked]
...
    ScheduledCallResponse:
      type: object
      description: Returned with HTTP 202 when `POST /calls` carries a future `scheduledAt`
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            scheduled:
              type: boolean
              example: true
            callbackId:
              type: string
              format: uuid
            scheduledAt:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
...
    CallbackVariables:
      type: object
      description: Up to 20 keys; values up to 300 characters. Available to the agent as `{{key}}`.
      maxProperties: 20
      additionalProperties:
        type: string
        maxLength: 300
      example: { first_name: "Priya", webinar_title: "GST filing for startups" }
    Callback:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contactName:
          type: string
          nullable: true
        contactPhone:
          type: string
          example: "+919876543210"
        reason:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
        timeZone:
          type: string
          example: Asia/Kolkata
        status:
          type: string
          enum: [pending, calling, completed, failed, cancelled]
        attempts:
          type: integer
        lastError:
          type: string
          nullable: true
        agentId:
          type: string
          format: uuid
          nullable: true
        agentName:
          type: string
          nullable: true
        resultCallId:
          type: string
          nullable: true
          description: The call placed for this callback — use it with `GET /calls/{id}`
        variables:
          $ref: '#/components/schemas/CallbackVariables'
        externalRef:
          type: string
          nullable: true
          maxLength: 120
        source:
          type: string
          enum: [agent, manual, api]
        createdAt:
          type: string
          format: date-time
    CallbackResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Callback'
        meta:
          $ref: '#/components/schemas/Meta'
    CallbackCancelResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            cancelled:
              type: boolean
              example: true
            callback:
              $ref: '#/components/schemas/Callback'
        meta:
          $ref: '#/components/schemas/Meta'
    CallbackListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Callback'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateCallbackRequest:
      type: object
      required: [agentId, contactPhone, scheduledAt]
      properties:
        agentId:
          type: string
          format: uuid
        contactPhone:
          type: string
          example: "+919876543210"
        scheduledAt:
          type: string
          format: date-time
          description: In the future, at most 60 days ahead. Include the offset (e.g. `+05:30`) or send UTC.
          example: "2026-09-29T11:00:00+05:30"
        contactName:
          type: string
        reason:
          type: string
          maxLength: 200
          description: Told to the agent so it knows what the call is about
        timeZone:
          type: string
          description: IANA zone shown in the dashboard
          default: Asia/Kolkata
        variables:
          $ref: '#/components/schemas/CallbackVariables'
        externalRef:
          type: string
          maxLength: 120
          description: Idempotency key — same ref returns the existing callback (HTTP 200)
...